refactor lcd_cam for only ISR
