+++ examples/lcd_cam_latch.rs
@@ -1,8 +1,9 @@
 //! Embassy "async" example of an ESP32-S3 driving a 64x64 HUB75 display using
-//! the LCD_CAM peripheral with a latch-based circuit.
+//! the LCD_CAM peripheral with a SmartLEDShield-style latch circuit.
 //!
-//! This example draws a simple gradient on the display and shows the refresh
-//! rate and render rate plus a simple counter.
+//! The ISR handles the entire BCM refresh loop — the async `swap()` method lets
+//! the display task exchange framebuffers without blocking. The display task
+//! runs on core 1 to keep core 0 free for application work.
 //!
 //! Following pins are used:
 //! - R1     => GPIO10
@@ -32,8 +33,6 @@ use defmt::info;
 use defmt_rtt as _;
 use embassy_executor::task;
 use embassy_executor::Spawner;
-use embassy_sync::blocking_mutex::raw::CriticalSectionRawMutex;
-use embassy_sync::signal::Signal;
 use embassy_time::Duration;
 use embassy_time::Instant;
 use embassy_time::Timer;
@@ -46,31 +45,24 @@ use embedded_graphics::text::Text;
 use embedded_graphics::Drawable;
 use esp_backtrace as _;
 use esp_hal::clock::CpuClock;
-use esp_hal::gpio::AnyPin;
 use esp_hal::gpio::Level;
 use esp_hal::gpio::Output;
 use esp_hal::gpio::OutputConfig;
 use esp_hal::gpio::Pin;
-use esp_hal::interrupt::software::SoftwareInterruptControl;
-use esp_hal::interrupt::Priority;
-use esp_hal::peripherals::LCD_CAM;
 use esp_hal::system::Stack;
 use esp_hal::time::Rate;
 use esp_hal::timer::timg::TimerGroup;
-use esp_hub75::framebuffer::compute_frame_count;
+use esp_hub75::framebuffer::bitplane::latched::DmaFrameBuffer;
 use esp_hub75::framebuffer::compute_rows;
-use esp_hub75::framebuffer::latched::DmaFrameBuffer;
 use esp_hub75::Color;
 use esp_hub75::Hub75;
 use esp_hub75::Hub75Pins8;
-use esp_rtos::embassy::InterruptExecutor;
 use heapless::String;
 #[cfg(feature = "log")]
 use log::info;
 
 esp_bootloader_esp_idf::esp_app_desc!();
 
-// When you are okay with using a nightly compiler it's better to use https://docs.rs/static_cell/2.1.0/static_cell/macro.make_static.html
 macro_rules! mk_static {
     ($t:ty,$val:expr) => {{
         static STATIC_CELL: static_cell::StaticCell<$t> = static_cell::StaticCell::new();
@@ -80,54 +72,33 @@ macro_rules! mk_static {
     }};
 }
 
-static REFRESH_RATE: AtomicU32 = AtomicU32::new(0);
 static RENDER_RATE: AtomicU32 = AtomicU32::new(0);
 static SIMPLE_COUNTER: AtomicU32 = AtomicU32::new(0);
 
 const ROWS: usize = 64;
 const COLS: usize = 64;
-const BITS: u8 = 4;
 const NROWS: usize = compute_rows(ROWS);
-const FRAME_COUNT: usize = compute_frame_count(BITS);
+const PLANES: usize = 6;
 
 const LINE1: i32 = ROWS as i32 - 1 - 14;
 const LINE2: i32 = ROWS as i32 - 1 - 7;
 const LINE3: i32 = ROWS as i32 - 1;
 const NBARS: i32 = ROWS as i32 / 8;
 
-type Hub75Type = Hub75<'static, esp_hal::Async>;
-type FBType = DmaFrameBuffer<ROWS, COLS, NROWS, BITS, FRAME_COUNT>;
-type FrameBufferExchange = Signal<CriticalSectionRawMutex, &'static mut FBType>;
-
-pub struct Hub75Peripherals<'d> {
-    pub lcd_cam: LCD_CAM<'d>,
-    pub dma_channel: esp_hal::peripherals::DMA_CH0<'d>,
-    pub red1: AnyPin<'d>,
-    pub grn1: AnyPin<'d>,
-    pub blu1: AnyPin<'d>,
-    pub red2: AnyPin<'d>,
-    pub grn2: AnyPin<'d>,
-    pub blu2: AnyPin<'d>,
-    pub blank: AnyPin<'d>,
-    pub clock: AnyPin<'d>,
-    pub latch: AnyPin<'d>,
-}
+type FBType = DmaFrameBuffer<NROWS, COLS, PLANES>;
 
 #[task]
-async fn display_task(
-    rx: &'static FrameBufferExchange,
-    tx: &'static FrameBufferExchange,
-    mut fb: &'static mut FBType,
-) {
+async fn display_task(hub75: Hub75<esp_hal::Async>, mut fb: &'static mut FBType) {
     info!("display_task: starting!");
     let fps_style = MonoTextStyleBuilder::new()
         .font(&FONT_5X7)
         .text_color(Color::YELLOW)
         .background_color(Color::BLACK)
         .build();
-    let mut count = 0u32;
-
+    let mut render_count = 0u32;
+    let mut refresh_count_start = hub75.frame_count();
     let mut start = Instant::now();
+    let mut refresh_rate = 0u32;
 
     loop {
         fb.erase();
@@ -150,11 +121,7 @@ async fn display_task(
 
         let mut buffer: String<64> = String::new();
 
-        fmt::write(
-            &mut buffer,
-            format_args!("Refresh: {:4}", REFRESH_RATE.load(Ordering::Relaxed)),
-        )
-        .unwrap();
+        fmt::write(&mut buffer, format_args!("Refresh: {:4}", refresh_rate)).unwrap();
         Text::with_alignment(
             buffer.as_str(),
             Point::new(0, LINE3),
@@ -170,7 +137,6 @@ async fn display_task(
             format_args!("Render: {:5}", RENDER_RATE.load(Ordering::Relaxed)),
         )
         .unwrap();
-
         Text::with_alignment(
             buffer.as_str(),
             Point::new(0, LINE2),
@@ -195,86 +161,17 @@ async fn display_task(
         .draw(fb)
         .unwrap();
 
-        // send the frame buffer to be rendered
-        tx.signal(fb);
+        let old_ptr = hub75.swap(fb).await.expect("DMA transfer failed");
+        fb = unsafe { &mut *(old_ptr as *mut FBType) };
 
-        // get the next frame buffer
-        fb = rx.wait().await;
-
-        // count up the rate we are rendering full buffer
-        count += 1;
+        render_count += 1;
         const FPS_INTERVAL: Duration = Duration::from_secs(1);
         if start.elapsed() > FPS_INTERVAL {
-            RENDER_RATE.store(count, Ordering::Relaxed);
-            count = 0;
-            start = Instant::now();
-        }
-    }
-}
-
-#[task]
-async fn hub75_task(
-    peripherals: Hub75Peripherals<'static>,
-    rx: &'static FrameBufferExchange,
-    tx: &'static FrameBufferExchange,
-    fb: &'static mut FBType,
-) {
-    info!("hub75_task: starting!");
-    let channel = peripherals.dma_channel;
-    let tx_descriptors = esp_hub75::hub75_dma_descriptors!(FBType);
-
-    let pins = Hub75Pins8 {
-        red1: peripherals.red1,
-        grn1: peripherals.grn1,
-        blu1: peripherals.blu1,
-        red2: peripherals.red2,
-        grn2: peripherals.grn2,
-        blu2: peripherals.blu2,
-        blank: peripherals.blank,
-        clock: peripherals.clock,
-        latch: peripherals.latch,
-    };
-
-    let mut hub75 = Hub75Type::new_async(
-        peripherals.lcd_cam,
-        pins,
-        channel,
-        tx_descriptors,
-        Rate::from_mhz(20),
-    )
-    .expect("failed to create Hub75!");
-
-    let mut count = 0u32;
-    let mut start = Instant::now();
-
-    // keep the frame buffer in an option so we can swap it
-    let mut fb = Some(fb);
-
-    loop {
-        // if there is a new buffer available, swap it and send the old one
-        if rx.signaled() {
-            let new_fb = rx.wait().await;
-            let old_fb = fb.replace(new_fb).unwrap();
-            tx.signal(old_fb);
-        }
-        if let Some(ref mut fb) = fb {
-            let mut xfer = hub75
-                .render(fb)
-                .map_err(|(e, _hub75)| e)
-                .expect("failed to start render!");
-            xfer.wait_for_done()
-                .await
-                .expect("render DMA transfer failed");
-            let (result, new_hub75) = xfer.wait();
-            hub75 = new_hub75;
-            result.expect("transfer failed");
-        }
-
-        count += 1;
-        const FPS_INTERVAL: Duration = Duration::from_secs(1);
-        if start.elapsed() > FPS_INTERVAL {
-            REFRESH_RATE.store(count, Ordering::Relaxed);
-            count = 0;
+            RENDER_RATE.store(render_count, Ordering::Relaxed);
+            let current_frame_count = hub75.frame_count();
+            refresh_rate = current_frame_count.wrapping_sub(refresh_count_start);
+            refresh_count_start = current_frame_count;
+            render_count = 0;
             start = Instant::now();
         }
     }
@@ -295,31 +192,29 @@ async fn main(_spawner: Spawner) {
     });
     info!("ROWS: {}", ROWS);
     info!("COLS: {}", COLS);
-    info!("BITS: {}", BITS);
-    info!("FRAME_COUNT: {}", FRAME_COUNT);
+    info!("PLANES: {}", PLANES);
+    info!("FB size: {}", core::mem::size_of::<FBType>());
     let peripherals = esp_hal::init(esp_hal::Config::default().with_cpu_clock(CpuClock::max()));
-    let sw_ints = SoftwareInterruptControl::new(peripherals.SW_INTERRUPT);
-    let software_interrupt = sw_ints.software_interrupt2;
 
     let timg0 = TimerGroup::new(peripherals.TIMG0);
+    let sw_ints =
+        esp_hal::interrupt::software::SoftwareInterruptControl::new(peripherals.SW_INTERRUPT);
 
     info!("init embassy");
     esp_rtos::start(timg0.timer0, sw_ints.software_interrupt0);
 
-    info!("init framebuffer exchange");
-    static TX: FrameBufferExchange = FrameBufferExchange::new();
-    static RX: FrameBufferExchange = FrameBufferExchange::new();
-
     info!("init framebuffers");
     let fb0 = mk_static!(FBType, FBType::new());
+    fb0.erase();
     let fb1 = mk_static!(FBType, FBType::new());
+    fb1.erase();
 
     info!("fb0: {:?}", fb0);
     info!("fb1: {:?}", fb1);
 
-    let hub75_peripherals = Hub75Peripherals {
-        lcd_cam: peripherals.LCD_CAM,
-        dma_channel: peripherals.DMA_CH0,
+    let tx_descriptors = esp_hub75::hub75_dma_descriptors!(FBType);
+
+    let pins = Hub75Pins8 {
         red1: peripherals.GPIO10.degrade(),
         grn1: peripherals.GPIO11.degrade(),
         blu1: peripherals.GPIO12.degrade(),
@@ -331,26 +226,24 @@ async fn main(_spawner: Spawner) {
         latch: peripherals.GPIO21.degrade(),
     };
 
-    // Configure OE_PWM pin as high output
     let _pwm_pin = Output::new(peripherals.GPIO48, Level::High, OutputConfig::default());
 
-    // run hub75 and display on second core
+    let hub75 = Hub75::new_async(
+        peripherals.LCD_CAM,
+        pins,
+        peripherals.DMA_CH0,
+        tx_descriptors,
+        Rate::from_mhz(30),
+    )
+    .expect("failed to create Hub75");
+    hub75.start(&*fb0).expect("failed to start Hub75");
+
     let cpu1_fnctn = {
         move || {
             use esp_rtos::embassy::Executor;
-            let hp_executor = mk_static!(
-                InterruptExecutor<2>,
-                InterruptExecutor::new(software_interrupt)
-            );
-            let high_pri_spawner = hp_executor.start(Priority::Priority3);
-
-            // hub75 runs as high priority task
-            high_pri_spawner.spawn(hub75_task(hub75_peripherals, &RX, &TX, fb1).unwrap());
-
-            let lp_executor = mk_static!(Executor, Executor::new());
-            // display task runs as low priority task
-            lp_executor.run(|spawner| {
-                spawner.spawn(display_task(&TX, &RX, fb0).unwrap());
+            let executor = mk_static!(Executor, Executor::new());
+            executor.run(|spawner| {
+                spawner.spawn(display_task(hub75, fb1).unwrap());
             });
         }
     };
