+++ examples/parl_io.rs
