+++ examples/parl_io_latch.rs
